--- src/components/layout/about.css
@@ -1,17 +1,0 @@
-.aboutimg {
-    width: 10vw;
-    height: auto;
-}
- .abouttxt {
-     font-size: 1.5vw;
-     width: 75vw;
-     height: auto;
- }
-
- .abouttitle {
-     font-size: 4vw;
- }
-
- .right {
-    direction:rtl;
- }
